made seperate files and folders for middleware, message routes, and user routes
